refactor: pull diagnostics (#585)

+++ crates/pgls_cli/src/execute/process_file/check.rs
@@ -54,7 +54,7 @@ pub(crate) fn check_with_guard<'ctx>(
                         .into_iter()
                         .map(Error::from)
                         .collect(),
-                    skipped_diagnostics: pull_diagnostics_result.skipped_diagnostics as u32,
+                    skipped_diagnostics: pull_diagnostics_result.skipped_diagnostics,
                 });
             }
